feat: create frontend
