+++ projects/client/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+    "extends": [
+        "next/core-web-vitals",
+        "next/typescript",
+        "plugin:storybook/recommended"
+    ]
+}
