+++ projects/client/.prettierrc.json
