+++ projects/client/README.md
@@ -1,0 +1,41 @@
+# フロントエンド
+
+## 概要
+
+これはフロントエンドのリポジトリです
+
+## 技術スタック
+
+-   Next.js
+-   vanilla-extract
+
+## 環境構築
+
+1. `pnpm`が入ってない人は`npm i -g pnpm`
+1. `pnpm i`でパッケージインストール
+
+## 開発環境起動
+
+```bash
+$ pnpm dev
+```
+
+## ビルド
+
+```bash
+$ pnpm build
+```
+
+## storybook 起動
+
+```bash
+$ pnpm storybook
+```
+
+## テスト実行
+
+```bash
+$ pnpm test
+```
+
+規約は後で書きます。
