+++ projects/client/next.config.ts
@@ -1,0 +1,10 @@
+import { createVanillaExtractPlugin } from "@vanilla-extract/next-plugin"
+import type { NextConfig } from "next"
+
+const withVanillaExtract = createVanillaExtractPlugin()
+
+const nextConfig: NextConfig = {
+    /* config options here */
+}
+
+export default withVanillaExtract(nextConfig)
