+++ projects/client/package.json
@@ -1,0 +1,51 @@
+{
+    "name": "client",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint",
+        "test": "vitest",
+        "storybook": "storybook dev -p 6006",
+        "build-storybook": "storybook build"
+    },
+    "dependencies": {
+        "@testing-library/jest-dom": "^6.6.3",
+        "@testing-library/react": "^16.0.1",
+        "@testing-library/user-event": "^14.5.2",
+        "@vanilla-extract/css": "^1.16.0",
+        "@vanilla-extract/next-plugin": "^2.4.6",
+        "@vanilla-extract/vite-plugin": "^4.0.17",
+        "@vanilla-extract/webpack-plugin": "^2.3.14",
+        "@vitejs/plugin-react": "^4.3.3",
+        "mini-css-extract-plugin": "^2.9.2",
+        "next": "15.0.3",
+        "path": "^0.12.7",
+        "react": "19.0.0-rc-66855b96-20241106",
+        "react-dom": "19.0.0-rc-66855b96-20241106",
+        "vitest": "^2.1.5"
+    },
+    "devDependencies": {
+        "@chromatic-com/storybook": "^3.2.2",
+        "@storybook/addon-essentials": "^8.4.5",
+        "@storybook/addon-interactions": "^8.4.5",
+        "@storybook/addon-onboarding": "^8.4.5",
+        "@storybook/addon-styling-webpack": "^1.0.1",
+        "@storybook/blocks": "^8.4.5",
+        "@storybook/nextjs": "^8.4.5",
+        "@storybook/react": "^8.4.5",
+        "@storybook/test": "^8.4.5",
+        "@types/node": "^20",
+        "@types/react": "^18",
+        "@types/react-dom": "^18",
+        "eslint": "^9",
+        "eslint-config-next": "15.0.3",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-storybook": "^0.11.1",
+        "jsdom": "^25.0.1",
+        "storybook": "^8.4.5",
+        "typescript": "^5"
+    }
+}
