+++ projects/client/src/app/favicon.ico
