+++ projects/client/src/app/layout.tsx
@@ -1,0 +1,33 @@
+import type { Metadata } from "next"
+import localFont from "next/font/local"
+import "./globals.css"
+
+const geistSans = localFont({
+    src: "./fonts/GeistVF.woff",
+    variable: "--font-geist-sans",
+    weight: "100 900"
+})
+const geistMono = localFont({
+    src: "./fonts/GeistMonoVF.woff",
+    variable: "--font-geist-mono",
+    weight: "100 900"
+})
+
+export const metadata: Metadata = {
+    title: "ハリーポッタアプリ",
+    description: "ハリーポッタアプリです"
+}
+
+export default function RootLayout({
+    children
+}: Readonly<{
+    children: React.ReactNode
+}>) {
+    return (
+        <html lang="ja">
+            <body className={`${geistSans.variable} ${geistMono.variable}`}>
+                {children}
+            </body>
+        </html>
+    )
+}
