+++ projects/client/src/app/page.tsx
@@ -1,0 +1,5 @@
+import { HomeComponent } from "@/feature/home"
+
+export default function Home() {
+    return <HomeComponent />
+}
