+++ projects/client/src/components/button-sample/style.css.ts
@@ -1,0 +1,15 @@
+import { style } from "@vanilla-extract/css"
+
+const styles = {
+    container: style({
+        width: 100,
+        height: 60,
+        backgroundColor: "red",
+        border: "none",
+        borderRadius: 8,
+        color: "white",
+        fontSize: 22
+    })
+}
+
+export default styles
