+++ projects/client/src/feature/home/index.ts
@@ -1,0 +1,1 @@
+export * from "./view/home"
