+++ projects/client/src/feature/home/view/home.tsx
@@ -1,0 +1,5 @@
+import { ButtonSample } from "@/components/button-sample"
+
+export const HomeComponent = () => {
+    return <ButtonSample>テストページhogehoge</ButtonSample>
+}
