--- projects/frontend/.eslintrc.json
@@ -1,28 +1,0 @@
-{
-    "extends": [
-        "next/core-web-vitals",
-        "plugin:storybook/recommended",
-        "prettier"
-    ],
-    "plugins": ["import", "react", "jsx-a11y", "unused-imports"],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": "latest",
-        "project": ["./tsconfig.json"],
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "sourceType": "module"
-    },
-    "rules": {
-        "import/order": [
-            "error",
-            {
-                "alphabetize": {
-                    "order": "asc"
-                }
-            }
-        ],
-        "unused-imports/no-unused-imports": "warn"
-    }
-}
