--- projects/frontend/README.md
@@ -1,59 +1,0 @@
-# フロントエンド
-
-## 1 概要
-
-こちらはフロントエンドのディレクトリです。
-
-## 2 開発環境
-
--   Next.js 14.2.3
--   react 18
--   vanilla-extract 1.15.2
--   storybook 8.1.3
--   vitest 1.6.0
-
-## 3 手順
-
-### 3.1 パッケージインストール
-
-```sh
-$pnpm install
-```
-
-### 3.2 各種実行
-
-#### 3.2.1 開発環境実行
-
-```sh
-$pnpm dev
-```
-
-#### 3.2.2 本番環境実行
-
-```sh
-$pnpm build
-```
-
-をした後に、
-
-```sh
-$pnpm start
-```
-
-#### 3.2.3 テスト実行
-
-```sh
-$pnpm test
-```
-
-#### 3.2.4 storybook 起動
-
-```sh
-$pnpm storybook
-```
-
-#### 3.2.5 `lint`確認
-
-```sh
-$pnpm lint
-```
