--- projects/frontend/next.config.mjs
@@ -1,8 +1,0 @@
-import { createVanillaExtractPlugin } from "@vanilla-extract/next-plugin";
-
-const withVanillaExtract = createVanillaExtractPlugin();
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {};
-
-export default withVanillaExtract(nextConfig);
