--- projects/frontend/package.json
@@ -1,56 +1,0 @@
-{
-    "name": "frontend",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint",
-        "storybook": "storybook dev -p 6006",
-        "build-storybook": "storybook build",
-        "test": "vitest --coverage"
-    },
-    "dependencies": {
-        "@vanilla-extract/css": "^1.15.2",
-        "next": "14.2.10",
-        "react": "^18",
-        "react-dom": "^18"
-    },
-    "devDependencies": {
-        "@chromatic-com/storybook": "^1.4.0",
-        "@storybook/addon-essentials": "^8.1.3",
-        "@storybook/addon-interactions": "^8.1.3",
-        "@storybook/addon-links": "^8.1.3",
-        "@storybook/addon-onboarding": "^8.1.3",
-        "@storybook/addon-styling-webpack": "^1.0.0",
-        "@storybook/blocks": "^8.1.3",
-        "@storybook/nextjs": "^8.1.3",
-        "@storybook/react": "^8.1.3",
-        "@storybook/test": "^8.1.3",
-        "@testing-library/jest-dom": "6.4.5",
-        "@testing-library/react": "^15.0.7",
-        "@testing-library/user-event": "^14.5.2",
-        "@types/node": "^20",
-        "@types/react": "^18",
-        "@types/react-dom": "^18",
-        "@vanilla-extract/next-plugin": "^2.4.1",
-        "@vanilla-extract/vite-plugin": "^4.0.10",
-        "@vanilla-extract/webpack-plugin": "^2.3.9",
-        "@vitejs/plugin-react": "^4.3.0",
-        "@vitest/coverage-v8": "^1.6.0",
-        "add": "^2.0.6",
-        "eslint": "^8",
-        "eslint-config-next": "14.2.3",
-        "eslint-config-prettier": "9.1.0",
-        "eslint-plugin-import": "^2.29.1",
-        "eslint-plugin-storybook": "^0.8.0",
-        "eslint-plugin-unused-imports": "^4.0.0",
-        "jsdom": "^24.1.0",
-        "mini-css-extract-plugin": "^2.9.0",
-        "pnpm": "^9.1.3",
-        "storybook": "^8.1.3",
-        "typescript": "^5",
-        "vitest": "^1.6.0"
-    }
-}
