--- projects/frontend/src/__test__/button/button.test.tsx
@@ -1,13 +1,0 @@
-import { render, screen } from "@testing-library/react"
-import { describe, expect, it } from "vitest"
-import { Button } from "@/components/sample"
-
-describe("<Button/>", () => {
-    it("要素が存在するか", () => {
-        render(<Button />)
-
-        const button = screen.getByRole("button")
-
-        expect(button).toBeInTheDocument()
-    })
-})
