--- projects/frontend/src/app/layout.tsx
@@ -1,22 +1,0 @@
-import type { Metadata } from "next"
-import { Inter } from "next/font/google"
-import "./globals.css"
-
-const inter = Inter({ subsets: ["latin"] })
-
-export const metadata: Metadata = {
-    title: "Harry Potter App",
-    description: "This is Harry Potter App."
-}
-
-export default function RootLayout({
-    children
-}: Readonly<{
-    children: React.ReactNode
-}>) {
-    return (
-        <html lang="ja">
-            <body className={inter.className}>{children}</body>
-        </html>
-    )
-}
