--- projects/frontend/src/components/sample/button.tsx
@@ -1,5 +1,0 @@
-import styles from "./style.css"
-
-export const Button = () => {
-    return <button className={styles.button}>Enter</button>
-}
