--- projects/frontend/src/components/sample/style.css.ts
@@ -1,14 +1,0 @@
-import { style } from "@vanilla-extract/css"
-
-const styles = {
-    button: style({
-        width: 100,
-        height: 30,
-        backgroundColor: "red",
-        color: "white",
-        border: "1px solid red",
-        borderRadius: 4
-    })
-}
-
-export default styles
